+++ packages/components/navigation/MiniModeNavigator.tsx
@@ -1,16 +1,16 @@
-import AsyncStorage from "@react-native-async-storage/async-storage";
 import {
   BottomTabNavigationProp,
   createBottomTabNavigator,
 } from "@react-navigation/bottom-tabs";
 import { RouteProp } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import React, { useEffect, useState } from "react";
+import React from "react";
 
 import { RootStackParamList } from "../../utils/navigation";
 import { neutral00, secondaryColor } from "../../utils/style/colors";
 import { TabBarIcon } from "../TabBarIcon";
 
+import { useOnboardedStatus } from "@/hooks/useOnboardStatus";
 import { FeedPostViewScreen } from "@/screens/FeedPostView/FeedPostViewScreen";
 import { NFTDetailScreen } from "@/screens/Marketplace/NFTDetailScreen";
 import AboutScreen from "@/screens/Mini/About/AboutScreen";
@@ -56,7 +56,6 @@ import { ImportWallet } from "@/screens/Wallet/Screens/ImportWallet";
 import NativeWallet from "@/screens/Wallet/Screens/NativeWallet";
 import { SuccessScreen } from "@/screens/Wallet/Screens/SucessScreen";
 import { ViewSeed } from "@/screens/Wallet/Screens/ViewSeed";
-import { safeParseJSON } from "@/utils/sanitize";
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
 const Tab = createBottomTabNavigator<RootStackParamList>();
@@ -96,24 +95,14 @@ const MainTab = () => {
   );
 };
 
-export const MiniModeNavigator: React.FC = () => {
-  const [isOnboarded, setIsOnboarded] = useState(false);
+export const MiniModeNavigator = () => {
+  const [isLoading, isOnboarded] = useOnboardedStatus();
 
-  useEffect(() => {
-    getOnboardedStatus();
-  }, []);
-
-  const getOnboardedStatus = async () => {
-    const onboarded = await AsyncStorage.getItem("ONBOARDED");
-    const data = safeParseJSON(onboarded as string) as boolean;
-    setIsOnboarded(data);
-  };
-
-  console.log(isOnboarded);
+  if (isLoading) return null;
 
   return (
     <Stack.Navigator
-      initialRouteName={!isOnboarded ? "ModeSelection" : "MiniTabs"}
+      initialRouteName={!isOnboarded ? "ModeSelection" : "NativeWallet"}
     >
       <Stack.Screen
         name="ModeSelection"
